package.json3

+++ src/components/blogs/Featured-post/index.tsx
@@ -12,18 +12,27 @@ export default function Featured({ lists }: { lists: Blog[] }) {
         router.refresh()
     },[])
 
-    async function handleDelete(id:number) {
+    async function handleDelete(id: number) {
       console.log(id);
-      
-      const res = await fetch(`/api/blog-post/delete-post?id=${id}`, {
-        method: "DELETE",
-        cache: "no-store",
-      });
-
-      const data = await res.json();
-
-      if (data && data.success) router.refresh();
+    
+      if (typeof window !== 'undefined') {
+        // Client-side logic
+        const res = await fetch(`/api/blog-post/delete-post?id=${id}`, {
+          method: "DELETE",
+          cache: "no-store",
+        });
+    
+        const data = await res.json();
+    
+        if (data && data.success) router.refresh();
+      } else {
+        // Server-side logic or skip during static generation
+        console.log("Running on the server or during static generation.");
+      }
     }
+    
+    
+    
 
   return (
     <section className="py-[120px]">
